GET,POST /broadcast (#5)
* [add] GET,POST /broadcat

* [add] Broadcast Test 코드 및 Paging 추가

+++ src/main/kotlin/com/broadcaster/api/entity/follow/FollowId.kt
@@ -3,6 +3,6 @@ package com.broadcaster.api.entity.follow
 import java.io.Serializable
 
 data class FollowId(
-    val broadcast:Int,
-    val users:Int
+    val broadcast:Long,
+    val users:Long
 ) : Serializable
